Move project text search into a worker.
